Refactor and improve Mailgun delivery time measurements at scale
Previously the exporter queried the Mailgun API for events in the last
minute and measured `delivery_time_seconds` by finding matching
`accepted` and `delivered` events. This assumes both events show up
within the last minute. However, we have seen delays of minutes
and even hours, since mail may be delayed by a variety of issues:
SMTP rate limits, busy servers, full mailboxes, etc.

This commit refactors the exporter and improves measurements on busy
Mailgun servers in a number of ways:

1. A time-limited hash table to track `acceptance` timestamps by a
unique message identifier. By default, a message will be tracked for
up to 10 hours before it is dropped.

2. The message ID in the SMTP header may not necessarily be unique if
in application generates it. We use the recipient as part of the
unique message identifier.

3. Increase the number of `delivery_time_seconds` buckets.

4. Drop the message ID from the `delivery_error` metric since the
cardinality of that label is too high.

5. Include the `domain` in all metrics.